Test,etc.: Test framework
The top level `Test` script sets up the build environment and runs the
tests. Any parameters passed to it are passed directly on to pytest.

`pactivate`[1] bootstraps the Python virtual environment and installs
in it any packages listed in `requirements.txt`.

`src/pytest_pt.py`[2] is a pytest plugin that allows us to use *.pt as
pytest files, so they are better distinguished from non-test *.py files.
The plugin is loaded by `src/conftest.py`.

`src/nothing.pt` is a single passing test to confirm that the build/test
framework is working. It should be removed when real tests are added.

`.gitignore` ignores the build directory, anything in a top-level tmp/
directory (for when developers have files for reference or whatever that
they don't want to commit), and __pycache__/ directories in case they've
not set $PYTHONPYCACHEPREFIX to put the python bytecode cache files
elsewhere.

[1]: https://github.com/cynic-net/pactivate
[2]: https://github.com/cynic-net/pytest_pt