src/binarytree.pt: Install binarytree package and test it
These tests confirm (and show) the behaviour of some of the aspects of the
binarytree library that we'd like to rely on, particularly with equality
tests.

The binarytree package emits a DeprecationWarning that will presumably be
fixed at some point, but for now we suppress it (with as much specificity
as possible, down to the line number) in pyproject.toml.