src/nodocF.py: src/formula.py with all documentation removed
I was just curious to see what this looked like, and I'm leaving
it here in case anybody else wants to have a look at it.

Note that there's nothing to keep this up to date with further
changes to src/formula.py. This should probably be removed soon
or, if we don't want to remove it, a tool built to generate this
file from formula.py.