src/formula.py: Change italic markup from Markdown to RST
Unlike Markdown, which can use _foo_ or *foo* for italic markup,
in reStructuredText you must use asterisks only. This emboldens
things a bit more than I care for in ASCII mode, but for the
moment we'll stick with the more standard (in the Python world)
RST in docstrings until we can make a more careful consideration
of what happens if we switch docstrings to Markdown.