Test: Print local documentation URL
This is stand-alone on a single line so that people with terminals
configured to select the entire line on triple-click or similar do not
have to edit out additional information at the beginning or the end.

This has not been tested on Windows, and likely has a small issue
there. the MINGW Bash-style paths and the Windows file: URL-style
paths are slightly different: the former does not use a colon for the
drive (`/c/Windows`) and the latter does (`///c:/Windows`). This is
easily enough tweaked by the user after pasting, though. Given that we
currently have no Windows users (that I'm aware of), fixing this is
not worth the added code complexity, but it's not hard to do if it
turns out later that we need it.