src/formula: Docstrings for node types, and fix a variable name
We had been assigning the node type Enum to `nodetype`, which conflicts
with the `nodetype()` function. (This is why the Enum wasn't appearing in
the docs.)  This didn't cause any issues because our only use of the enum
object was immediately after its definition to define the `VAR` etc.
constants. But we rename the first `nodetype` to `_nt` anyway, to help
avoid confusion.

And while we're here, add docstrings to the `VAR` etc. constants so
the documentation looks nicer to people not familiar with the code.
(Developers of this system would find it fairly obvious what they mean,
especially since they're mentioned as the return values of nodetype().)