src/formula: Make nodify() a private method
I can't see this being used by those using this as library, and it was
quite annoyingly appearing between the node type constants and the
nodetype() function in the documentation. (The function is in the correct
place for readers of the code, since it's used only by the constructor
directly above it.)

A good argument could be made that  the node type constants should be moved
further down in the file, but I prefer to have them up at the top of the
class, right after the class docstring that has general discussion of what
an `F` is. (Not that they stay there in the docs; pdoc always puts the
constructor immediately after the class docstring.)