src/formula: Make __eq__() a public method (included in generated docs)
This _really_ wants to be in the generated documentation since it
defines what equality for formulae objects means (and the current
caveats), which is kind of important.

This is done by adding an `@public` tag to the docstring, which is
recognised by pdoc. (pdoc3 does not recognise this; I'm not sure if
Sphinx does.)