src/pytest_pt.py: Replace with the PyPI package
This allows us to remove both src/pytest_pt.py and src/conftest.py;
the former is supplied by the pytest_pt distribution package and the
latter is no longer needed to activate it. Instead we now add '-p
pytest_pt' to the pytest.addopts setting in pyproject.toml.

The removal of this also means that src/ is no longer in our Python
sys.path, so we also need to add pythonpath = ['src'] to the [pytest]
section of our pyproject.toml.

As a bonus, this means that pytest_pt.py is no longer checked by mypy
if you happen to run it, so all those errors disappear. (Though the
source project should be running this and fixing those errors, of
course.)

doc/explainer has been updated; the "Source Files: Infrastructure"
section is now entirely replaced by a "Dependencies" section. This
still covers just pytest_pt but notes that other dependencies are
covered where they're used, and those parts have been checked and,
for pdoc, updated slightly.