doc/explainer: Fix a broken reference that was breaking many references
For qutie a while the [`.build/hdoc/`] reference definition was
missing a colon after the right bracket, which broke not only it, but
everything after it in the paragraph. (I'm not clear on whether this
is a bug in the GitHub (and [grip]) markdown parser, a bug in how
Markdown works, or just an artifact of Markdown syntax.)

The annoying thing here is that my syntax highlighting was displaying
a different colour, indicating a problem, but I ignored this rather
than tracking down what was wrong.

But what I really need here (and in most of my other projects, for
that matter) is a Markdown link checker that confirms that all
references in the text have a valid reference definition, and that the
links in the reference definition (whether to local files or the web)
are valid. That's non-trivial, so I won't deal with it now, but it's
something I clearly need to find or build.

[grip]: https://pypi.org/project/grip/