Test: Make this work from subdirs, e.g., `../Test foo.py`
This lets users who happen to be in a subdir, e.g., src/, pass paths
relative to that dir (usually generated via command-line completion) to
pytest, e.g:

    cd src/
    ../Test ./binarytree.py

This is valid only in subdirs of the project dir; if Test is run from
outside the project dir you need to specify paths relative to the project
dir.

This is slightly tricky, due to pytest ignoring its `testpaths`
configuration parameter if it's not started from the pytest rootdir: If you
run just `../Test` from doc/ it will search the CWD downward and find no
test files. I don't know of any way to fix this, but it can be useful
behaviour, e.g.:

    cd tmp/
    vi stuff_I_want_to_play_with.pt # add some tests
    ../Test

will discover just the stuff you're playing with under tmp/.