Test,pyproject.toml: Add mypy support
As well as adding mypy to requirements.txt, configuring it in
pyproject.toml and running it (before pytest!) in Test, this required
stub generation for the binarytree library, which without this
generates a 'missing library stubs or py.typed marker' complaint.

The stubs were generated with `stubgen -o mypy-stubs -p binarytree`.
The only edit required was to add a `# type: ignore` annotation to the
graphviz  import in mypy-stubs/binarytree/__init__.pyi. (I looked
around for stubs for graphviz but didn't find any; it doesn't seem
worth generating them for something currently only used internally by
`binarytree`.)

There's currently no way of passing arguments through Test to mypy;
there's further discussion of this in the comment in Test. And of
course it can always be run manually with one of:

    .build/virtualenv/bin/mypy ...
    pae . mypy ...                     [1]

We do not currently type check our test (*.pt) files. Testing
a single file and its tests with `mypy src/f*` seems to work
fine, but mypy has some issues with module names when testing
`src/*.p[yt]`. As well as attacking this directly, there's also
the option of looking at pytest-mypy[2].

[1]: https://github.com/cynic-net/pactivate/blob/main/doc/pae.md
[2]: https://pypi.org/project/pytest-mypy/