src/formula.py: Replace 'φ', 'A', etc. with just φ, A, etc.
We do this by defining Python variables φ, A, etc. that are
instances of F('φ'), F('A'), etc. This gets rid of having to
use quotes all over the place when creating new formulae.

We didn't previously allow passing in a formula as the `vc` (variable or
connective) to F(). The changes to allow this led us to rename F.nodetype()
to F.valtype(), since it's no longer about just nodes (not that it ever
was; it also accepted strings). And of course this implied that the
NodeValue type should be changed to just Value, which is done.

Various other changes started to surface areas where we're not quite as
clear as we could be, which have been commented, and also surfaced the
tension between a formula and a schema, which we've never addressed (but
which no doubt will come out further as we move on to proof steps and
proofs).

This unfortunately produces a diff that's rather hard to read,
but in this particular case I feel it's not worth the time to go
back and break it down into some smaller commits.