src/formula: Rename F,NO → Fm,No
Now that we've introduced "constants" for single-variable formulae as the
variable name itself (A, φ, etc.), thing get a bit awkward with calling
the formula constructor `F`, because that looks like a variable name.

So, new naming scheme: All formula constructors are two letters, an upper
case followed by a lower case, giving us `Fm` as the generic formula
constructor and `No` as the negation formula constructor.

And of course now we can add constants for the formula 'E' and 'F' now.