src/formula: remove binarytree; it was not pulling its weight
The binarytree module was not doing enough to justify using it here.
It gave us:
1. A binary tree structure, which we can do just as well ourselves
   with "value," "left" and "right" attributes directly on the Fm
   formula class. (Actually better, because we no longer need to
   extract the tree to recurse it because we _are_ the tree.)
2. A nice tree display, which was useful at the very start, but is no
   longer necessary that now that we have a __str__() that can display
   the formulae directly. (Working out the binary tree structure from
   that is trivial.)
3. Restriction of the type of the value to an int, float or str, which
   is quite a bit broader than we wanted anyway.
4. build() and build2() constructors, which we thought we might need,
   but never did because we developed our own construction technique
   that works better for us.
5. The clone() method, which we used to help avoid aliasing, but we
   can live without that if we make the Fm objects closer to immutable.

This cleans up and simplifies a fair amount of code now that we an `Fm`
_is_ a node rather than holding one. We do lose the ability to easily clone
trees, but that was used to avoid aliasing, and so we replace it with
read-only "public" attributes (value, type, left, right) that should catch
inadvertent instances of someone modifying an `Fm` instance. (Possibly we
will need to bring back the cloning if someone develops a need to clone
formulae, but I currently don't see any need.)

Note that we've not yet removed binarytree etc. from the project
requirements; this will happen in a later commit for easier review.