requirements,etc.: Remove `binarytree` requirement
This is no longer used by src/formula.*, leaving just src/binarytree.pt,
which tests we no longer need since we're not using binarytree.

We also no longer need the warnings filter for the binarytree warning in
pyproject.toml, but we leave it there anyway because it's a useful example
of how to properly deal with warnings (i.e., by narrowing them down to the
most specific filter possible) in pytest.