formula: Cache str(), repr() and vars() values
These are used pretty heavily and are expensive to calculate for
large formulae.

Normally I wouldn't do this without profiling to see if this is really
an issue, but it seemed worth demonstrating the technique here.
Though, now that I look at it, it seems that one could abstract this
out into a decorator method.

This could also be done with @lru_cache, but that in itself is more
complex and doesn't bind the lifetime of the cached values to the
lifetime of the objects.