proof: Revert to using `i in range(…)`; it's no slower
This essentially reverts commit f34810e. Based on a suggestion from
someone I had thought that the (0 < hyp < i) test would be faster.
Then I looked up the code in the Python source, which is all in C, and
thought it would be slower. Then I actually benchmarked it (in the
last commit) and it turns out to be exactly the same.

So I'm reverting this not because it makes any performance difference,
but just because I find the older version easier to read.