Test: Remove top-level `src` package from generated documentation
`pdoc` takes paths to "modules" (files and directories) and recurses
down them. We had mistakenly provided `src/` making it a top-level
module (a namespace module, since it has no __init__.py), but what we
really wanted was all the things under src/ to be the top-level
modules, just as they are imported into python.

If given an explicit `foo.pt` file it will consider that to be a `foo`
module, so we enable the Bash extglob option to let us use a glob that
excludes just those. (When recursing directories it seems to include
only subdirectories and *.py files, so we need not worry about
excluding *.pt except at the top level.)