src/Documentation: Documentation independent of Python modules
We have considerable documentation that isn't associated with any
particular Python module, and is certainly not API documentation, but
is important for understanding the project domain. This has to date
been available only under doc/, and is not in the generated
documentation, so could only be viewed in the repo itself or (in
rendered format) on the GitHub project pages for this repo.

This is an attempt to bring that documentation into the generated
documentation. Here we mainly just create new Python "docstring-only"
modules and use a reStructuredText `include` directive to copy in the
doc/*.md files. This is not too bad, but has several annoyances:

* The Markdown parsing[1] is slightly different from the
  GitHub-compatible one we normally use in doc/*.md files. Most
  obviously, it's enabled the markdown2 package's code-friendly[2]
  "extra" that doesn't allow underscore for _italics_ and the like.

* Many essentially spurious module names are added to the module list
  at the left. (Having the ability to expand and collapse the tree
  would help here.)

Some of this can probably be worked around with custom pdoc templates,
but that appears to involve copying a lot of code. Using a more
sophisticated documentation generator tool such as Sphinx might offer
the ability to tweak things more easily.

At the least, this is a useful experiment to learn more about
generating sites that include documentation beyond just the API docs.

[1]: https://pdoc.dev/docs/pdoc.html#markdown-support
[2]: https://github.com/trentm/python-markdown2/wiki/code-friendly